fix: clean up utils files and imports
