--- packages/config/src/utils.ts
@@ -1,18 +1,0 @@
-/**
- * Convert a string or number to an integer
- * @param value - The value to convert
- * @returns The integer value or undefined if the value is not a number
- * @throws If the value is not a number
- */
-export function toInt(value: number | string): number {
-    if (typeof value === "number") {
-        return value
-    }
-    const parsed = parseInt(value, 10)
-    if (isNaN(parsed)) {
-        throw new Error(
-            `${value} is not a number \n \t\t\tHINT: likely an invalid environment variable`
-        )
-    }
-    return parsed
-}
