+++ packages/logger/tsconfig.json
@@ -3,9 +3,9 @@
     "compilerOptions": {
         "outDir": "dist",
         "rootDir": "src",
+        "baseUrl": "src/",
         "moduleResolution": "bundler",
         "module": "ESNext",
-        "baseUrl": "src/",
         "paths": {
             "@/*": ["./*"]
         }
